Allow for injecting username and passwords in AMQ
- Create a new entrypoint script
- Modify Dockerfile to add and call entrypoint
- Removed hardcoded username and password from activemq.xml
- Insert placeholders in activemq.xml for username and password

This setup willl use the default username of 'messaging' and the
password of 'moo' if no env vars are passed.

If SPRING_ACTIVEMQ_USER and SPRING_ACTIVEMQ_PASSWORD are
passed, these values will be used.  If not, it will look for
ACTIVEMQ_USER and ACTIVEMQ_PASSWORD and use those.

Otherwise, the default values are used.